toolbox-companies/pyproject.toml
[project]
name = "toolbox-companies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-crc32c>=1.6.0",
    "langchain-community>=0.3.19",
    "langchain-google-vertexai>=2.0.13",
    "langgraph>=0.2.72",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.1",
    "toolbox-langchain>=0.1.0",
    "typing>=3.10.0.0",
]
